requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

django
boto3
botocore
requests
psycopg2-binary
djangorestframework
dynamodb_json
django-ninja